PWX-34145: Extracting getKvdbMap to Utils(#1316)
* Exposing Portworx struct

* Syntax Correction

* extracted getKvdbMap to util

* Exposing Portworx struct

* Syntax Correction

* extracted getKvdbMap to util

* Updated the requested changes

* Minor change

* resolved fmtcheck and conficts
