+++ drivers/storage/portworx/status.go
@@ -2,7 +2,6 @@ package portworx
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"reflect"
 	"strings"
@@ -26,11 +25,6 @@ import (
 	kvdb_api "github.com/portworx/kvdb/api/bootstrap"
 	coreops "github.com/portworx/sched-ops/k8s/core"
 	operatorops "github.com/portworx/sched-ops/k8s/operator"
-)
-
-const (
-	// pxEntriesKey is key which holds all the bootstrap entries
-	pxEntriesKey = "px-entries"
 )
 
 func (p *portworx) UpdateStorageClusterStatus(
@@ -382,32 +376,7 @@ func getStorageClusterState(
 func (p *portworx) getKvdbMap(
 	cluster *corev1.StorageCluster,
 ) map[string]*kvdb_api.BootstrapEntry {
-	// If cluster is running internal kvdb, get current bootstrap nodes
-	kvdbNodeMap := make(map[string]*kvdb_api.BootstrapEntry)
-	if cluster.Spec.Kvdb != nil && cluster.Spec.Kvdb.Internal {
-		clusterID := pxutil.GetClusterID(cluster)
-		strippedClusterName := strings.ToLower(pxutil.ConfigMapNameRegex.ReplaceAllString(clusterID, ""))
-		cmName := fmt.Sprintf("%s%s", pxutil.InternalEtcdConfigMapPrefix, strippedClusterName)
-
-		cm := &v1.ConfigMap{}
-		err := p.k8sClient.Get(context.TODO(), types.NamespacedName{
-			Name:      cmName,
-			Namespace: bootstrapCloudDriveNamespace,
-		}, cm)
-		if err != nil {
-			logrus.Warnf("failed to get internal kvdb bootstrap config map: %v", err)
-		}
-
-		// Get the bootstrap entries
-		entriesBlob, ok := cm.Data[pxEntriesKey]
-		if ok {
-			kvdbNodeMap, err = blobToBootstrapEntries([]byte(entriesBlob))
-			if err != nil {
-				logrus.Warnf("failed to get internal kvdb bootstrap config map: %v", err)
-			}
-		}
-	}
-	return kvdbNodeMap
+	return pxutil.GetKvdbMap(p.k8sClient, cluster)
 }
 
 func (p *portworx) updateStorageNodes(
@@ -881,21 +850,4 @@ func getStorageNodePhase(status *corev1.NodeStatus) string {
 		return string(nodeInitCondition.Status)
 	}
 	return string(nodeStateCondition.Status)
-}
-
-func blobToBootstrapEntries(
-	entriesBlob []byte,
-) (map[string]*kvdb_api.BootstrapEntry, error) {
-
-	var bEntries []*kvdb_api.BootstrapEntry
-	if err := json.Unmarshal(entriesBlob, &bEntries); err != nil {
-		return nil, err
-	}
-
-	// return as a map by ID to facilitate callers
-	retMap := make(map[string]*kvdb_api.BootstrapEntry)
-	for _, e := range bEntries {
-		retMap[e.ID] = e
-	}
-	return retMap, nil
 }
